creature: move Creature_EnsureHabitat to trx
