+++ src/tr1/game/creature.h
@@ -9,24 +9,9 @@
 #define CREATURE_SHOOT_RANGE SQUARE(WALL_L * 7) // = 51380224
 #define CREATURE_MISS_CHANCE 0x2000
 
-typedef struct {
-    struct {
-        GAME_OBJECT_ID id;
-        int16_t active_anim;
-        int16_t death_anim;
-        int16_t death_state;
-    } land;
-    struct {
-        GAME_OBJECT_ID id;
-        int16_t active_anim;
-    } water;
-} HYBRID_INFO;
-
 bool Creature_CheckBaddieOverlap(int16_t item_num);
 bool Creature_CanTargetEnemy(ITEM *item, AI_INFO *info);
 bool Creature_ShootAtLara(
     ITEM *item, int32_t distance, BITE *gun, int16_t extra_rotation,
     int16_t damage);
-bool Creature_EnsureHabitat(
-    int16_t item_num, int32_t *wh, const HYBRID_INFO *info);
 bool Creature_IsBoss(int16_t item_num);
